Tictactoe 클래스 역할 분리
